refactor(schedule): reorganize components into Day/Week/Shared modules
Extract and reorganize CourseSchedule components into logical directories
based on their usage patterns. Day-specific components moved to Day/,
Week-specific to Week/, and shared components to Shared/.
