+++ dauphin/View/CourseSchedule/CourseScheduleView.swift
@@ -18,7 +18,7 @@ struct CourseScheduleView: View {
       Group {
         if authViewModel.isLoggedIn {
           if horizontalSizeClass == .compact {
-            CourseScheduleByDayView(courseViewModel: viewModel, authViewModel: authViewModel)
+            DayScheduleView(courseViewModel: viewModel, authViewModel: authViewModel)
               .refreshable {
                 if !authViewModel.ssoStuNo.isEmpty {
                   Task {
@@ -38,7 +38,7 @@ struct CourseScheduleView: View {
                 }
               }
           } else {
-            CourseScheduleByWeekView(courseViewModel: viewModel)
+            WeekScheduleView(courseViewModel: viewModel)
               .padding(
                 [.horizontal], 15
               )
@@ -68,42 +68,10 @@ struct CourseScheduleView: View {
       }
 
       // Floating notification overlay
-      VStack {
-        if let message = viewModel.cacheUpdateMessage {
-          HStack {
-            if viewModel.isUpdatingCache {
-              ProgressView()
-                .progressViewStyle(CircularProgressViewStyle())
-                .scaleEffect(0.8)
-                .padding(.trailing, 4)
-            } else {
-              Image(systemName: "checkmark.circle.fill")
-                .foregroundColor(.green)
-                .padding(.trailing, 4)
-            }
-            Text(message)
-              .font(.footnote)
-              .fontWeight(.medium)
-          }
-          .padding(.horizontal, 16)
-          .padding(.vertical, 10)
-          .background(
-            Capsule()
-              .fill(.regularMaterial)
-              .shadow(color: Color.black.opacity(0.1), radius: 4, x: 0, y: 2)
-          )
-          .transition(
-            .asymmetric(
-              insertion: .move(edge: .top).combined(with: .opacity),
-              removal: .move(edge: .top).combined(with: .opacity)
-            ))
-        }
-        Spacer()
-      }
-      .padding(.top, 50)
-      .animation(
-        .spring(response: 0.35, dampingFraction: 0.85, blendDuration: 0),
-        value: viewModel.cacheUpdateMessage)
+      CacheUpdateNotificationView(
+        message: viewModel.cacheUpdateMessage,
+        isUpdating: viewModel.isUpdatingCache
+      )
     }
   }
 }
