+++ dauphin/View/CourseSchedule/Shared/CourseDetailView.swift
