+++ dauphin/View/CourseSchedule/Week/CoursePosition.swift
@@ -1,0 +1,14 @@
+//
+//  CoursePosition.swift
+//  dauphin
+//
+//  Model for course positioning in timeline view
+//
+
+import Foundation
+
+struct CoursePosition {
+    let course: Course
+    let column: Int
+    let totalColumns: Int
+}
