+++ dauphin/View/CourseSchedule/Week/TimelineView.swift
@@ -1,5 +1,5 @@
 //
-//  SingleTimelineView.swift
+//  TimelineView.swift
 //  dauphin
 //
 //  Created by \u8b19 on 12/13/24.
@@ -7,7 +7,7 @@
 
 import SwiftUI
 
-struct SingleTimeline: View {
+struct TimelineView: View {
   @Binding var courses: [Course]
   var onCourseTap: ((Course) -> Void)? = nil
   var overlapGap: CGFloat = -4
@@ -16,13 +16,6 @@ struct SingleTimeline: View {
     bySettingHour: 8, minute: 0, second: 0, of: Calendar.current.startOfDay(for: Date()))!
   let end = Calendar.current.date(
     bySettingHour: 22, minute: 0, second: 0, of: Calendar.current.startOfDay(for: Date()))!
-
-  // Structure to hold course positioning information
-  struct CoursePosition {
-    let course: Course
-    let column: Int
-    let totalColumns: Int
-  }
 
   // Calculate overlapping groups and assign columns
   private var positionedCourses: [CoursePosition] {
@@ -200,77 +193,10 @@ struct SingleTimeline: View {
   }
 }
 
-struct TimeSlotGrid: View {
-  let numberOfSlots: Int
-  let totalHeight: CGFloat
-
-  var body: some View {
-    VStack(spacing: 0) {
-      ForEach(0..<numberOfSlots, id: \.self) { index in
-        Rectangle()
-          .stroke(Color.gray.opacity(0.4), lineWidth: 0.3)
-          .frame(height: totalHeight / CGFloat(numberOfSlots))
-      }
-    }
-  }
-}
-
-struct CourseView: View {
-  let course: Course
-  let height: CGFloat
-  let yOffset: CGFloat
-
-  private var courseColor: Color {
-    CourseColors.color(for: course.name)
-  }
-
-  var body: some View {
-    RoundedRectangle(cornerRadius: 8)
-      .fill(courseColor.opacity(0.85))
-      .frame(height: max(height - 4, 20))  // Fixed 4pt gap for even spacing
-      .overlay(
-        RoundedRectangle(cornerRadius: 8)
-          .strokeBorder(courseColor, lineWidth: 1.5)
-      )
-      .overlay(
-        VStack(alignment: .leading, spacing: 3) {
-          HStack {
-            Text(course.name)
-              .font(.system(size: 11, weight: .semibold))
-              .foregroundColor(.white)
-              .lineLimit(2)
-            Spacer()
-          }
-          HStack(spacing: 3) {
-            Image(systemName: "location.circle.fill")
-              .font(.system(size: 9))
-              .foregroundColor(.white.opacity(0.9))
-            Text(course.room)
-              .font(.system(size: 10))
-              .foregroundColor(.white.opacity(0.95))
-          }
-          HStack(spacing: 3) {
-            Image(systemName: "number.circle.fill")
-              .font(.system(size: 9))
-              .foregroundColor(.white.opacity(0.9))
-            Text(course.stdNo)
-              .font(.system(size: 10))
-              .foregroundColor(.white.opacity(0.95))
-          }
-        }
-        .padding(8),
-        alignment: .topLeading
-      )
-      .shadow(color: courseColor.opacity(0.3), radius: 2, x: 0, y: 1)
-      .offset(y: yOffset)
-      .padding(.horizontal, 4)
-  }
-}
-
 #Preview {
   let courseViewModel = CourseViewModel(mockData: mockData)
   ScrollView {
-    SingleTimeline(
+    TimelineView(
       courses: Binding(
         get: { courseViewModel.weekCourses.filter { $0.weekday == 1 } },
         set: { newValue in
